requirements.txt
openai>=1.0.0
sentence-transformers>=2.2.0
numpy>=1.24.0
flask>=3.0.0
python-dotenv>=1.0.0
pytest>=7.0.0
